+++ internal/av1/tile/mv_test.go
@@ -380,7 +380,7 @@ func FuzzReadMotionVector(f *testing.F) {
 			t.Fatal(err)
 		}
 		var state DecodeState
-		if err := state.Reset(payload, Job{Offset: 0, Size: len(payload)}, DecodeOptions{}); err != nil {
+		if err := state.Reset(payload, Job{Offset: 0, Size: uint32(len(payload))}, DecodeOptions{}); err != nil {
 			t.Fatal(err)
 		}
 		_, residual, err := state.ReadMotionVector(&cdfs, motion.Vector{Row: row, Col: col}, precision)
