+++ internal/av1/tile/partition_test.go
@@ -263,7 +263,7 @@ func FuzzReadPartition(f *testing.F) {
 			t.Fatal(err)
 		}
 		var state DecodeState
-		if err := state.Reset(payload, Job{Offset: 0, Size: len(payload)}, DecodeOptions{}); err != nil {
+		if err := state.Reset(payload, Job{Offset: 0, Size: uint32(len(payload))}, DecodeOptions{}); err != nil {
 			t.Fatal(err)
 		}
 		partition, err := state.ReadPartition(&cdfs, level, ctx, haveH, haveV)
